+++ src/lib/components/export/index.ts
@@ -1,3 +1,2 @@
 export { default as ExportModal } from './ExportModal.svelte'
-export { default as ExportProgress } from './ExportProgress.svelte'
 export { default as ExportFailureModal } from './ExportFailureModal.svelte'